consolidate some functions
